requirements.txt
torch==1.9.1
torchvision==0.10.1
matplotlib==3.3.2
numpy==1.19.5
pillow==8.4.0
python==3.6.13
pyyaml==6.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.3
habitat-sim==0.1.7